src/scraping_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
